feat: Enhance OpenAI integration with language detection pypi library and model name update.
now instead of the llm making the distinction of whether the language is not in spanish, the library language-detect will decide using the users message of at least 2 words.

model was update to gpt-4o as its mini version was failing to translate the documents given. lang detect is a remediation for this, if the language is already in spanish, there will not be a need to send an Openai call.